abstrai método de autorizacao para classe própria
